requirements.txt
boto3==1.26.0
Pillow==9.3.0